backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["MasaruFukazawa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
psycopg2 = "^2.9.9"
psycopg = "^3.1.18"
djangorestframework = "^3.15.1"
black = "^24.4.2"
isort = "^5.13.2"
bandit = "^1.7.8"
flake8 = "^7.0.0"
djangorestframework-simplejwt = "^5.3.1"
timedelta = "^2020.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
